backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
motor==3.3.2
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
python-jose==3.3.0
httpx==0.26.0
python-multipart==0.0.9 